Switch to LaunchQueueService
